// plugwerk-server/plugwerk-server-frontend/src/api/generated/model/plugin-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Plugwerk API
 * Plugin marketplace for the Java/PF4J ecosystem
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface PluginDto
 */
export interface PluginDto {
    /**
     * 
     * @type {string}
     * @memberof PluginDto
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof PluginDto
     */
    'pluginId': string;
    /**
     * 
     * @type {string}
     * @memberof PluginDto
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof PluginDto
     */
    'description'?: string;
    /**
     * 
     * @type {string}
     * @memberof PluginDto
     */
    'author'?: string;
    /**
     * 
     * @type {string}
     * @memberof PluginDto
     */
    'license'?: string;
    /**
     * 
     * @type {string}
     * @memberof PluginDto
     */
    'namespace'?: string;
    /**
     * 
     * @type {string}
     * @memberof PluginDto
     */
    'status': PluginDtoStatusEnum;
    /**
     * 
     * @type {Array<string>}
     * @memberof PluginDto
     */
    'categories'?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof PluginDto
     */
    'tags'?: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof PluginDto
     */
    'latestVersion'?: string;
    /**
     * 
     * @type {number}
     * @memberof PluginDto
     */
    'downloadCount'?: number;
    /**
     * 
     * @type {string}
     * @memberof PluginDto
     */
    'icon'?: string;
    /**
     * 
     * @type {string}
     * @memberof PluginDto
     */
    'homepage'?: string;
    /**
     * 
     * @type {string}
     * @memberof PluginDto
     */
    'repository'?: string;
    /**
     * 
     * @type {string}
     * @memberof PluginDto
     */
    'createdAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof PluginDto
     */
    'updatedAt'?: string;
}

export const PluginDtoStatusEnum = {
    Active: 'active',
    Suspended: 'suspended',
    Archived: 'archived'
} as const;

export type PluginDtoStatusEnum = typeof PluginDtoStatusEnum[keyof typeof PluginDtoStatusEnum];